+++ test/Unchase.Swashbuckle.AspNetCore.Extensions.Tests/Models/EnumWithoutXmlComments.cs
@@ -1,0 +1,20 @@
+﻿namespace Unchase.Swashbuckle.AspNetCore.Extensions.Tests.Models
+{
+    /// <summary>
+    /// Enum without xml-comments
+    /// </summary>
+    public enum EnumWithoutXmlComments
+    {
+        /// <summary>
+        /// This is first
+        /// </summary>
+        First,
+
+        /// <summary>
+        /// This is second
+        /// </summary>
+        Second,
+
+        Third
+    }
+}